+++ BE/src/main/java/com/issuetracker/service/LabelService.java
@@ -1,5 +1,6 @@
 package com.issuetracker.service;
 
+import com.issuetracker.exception.LabelNotFoundException;
 import com.issuetracker.web.dto.response.LabelDTO;
 import com.issuetracker.domain.issue.Issue;
 import com.issuetracker.domain.label.Label;
@@ -53,9 +54,7 @@ public class LabelService {
     }
 
     private Label findLabelById(Long id) {
-        return labelRepository.findById(id).orElseThrow(
-                () -> new ElementNotFoundException("Cannot find label by given id.")
-        );
+        return labelRepository.findById(id).orElseThrow(LabelNotFoundException::new);
     }
 
     public List<LabelDTO> labelsToLabelDTOs(Issue issue) {
