+++ BE/src/main/java/com/issuetracker/web/IssueController.java
@@ -1,7 +1,8 @@
 package com.issuetracker.web;
 
 import com.issuetracker.auth.annotation.LoginRequired;
-import com.issuetracker.service.IssueService;
+import com.issuetracker.service.IssueCommandService;
+import com.issuetracker.service.IssueQueryService;
 import com.issuetracker.web.dto.reqeust.*;
 import com.issuetracker.web.dto.response.*;
 import com.issuetracker.auth.annotation.UserId;
@@ -15,26 +16,27 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class IssueController {
 
-    private final IssueService issueService;
+    private final IssueQueryService issueQueryService;
+    private final IssueCommandService issueCommandService;
     private final Logger logger = LoggerFactory.getLogger(IssueController.class);
 
     @GetMapping
     public IssuesResponseDTO view(SearchRequestDTO requestDTO) {
         logger.debug("모든 이슈 조회");
-        return issueService.getIssues(requestDTO);
+        return issueQueryService.getIssues(requestDTO);
     }
 
     @PatchMapping
     public void changeStatus(@RequestBody IssueNumbersRequestDTO issueNumbersRequestDTO, @RequestParam String status) {
         logger.debug("이슈 닫기 or 열기");
         logger.debug("issue 상태 변경 확인: {}", issueNumbersRequestDTO.toString());
-        issueService.changeIssueStatus(issueNumbersRequestDTO, status);
+        issueCommandService.changeIssueStatus(issueNumbersRequestDTO, status);
     }
 
     @GetMapping("/form")
     public IssueFormResponseDTO viewForm() {
         logger.debug("이슈 생성 페이지 요청");
-        return issueService.getIssueForm();
+        return issueQueryService.getIssueForm();
     }
 
     @LoginRequired
@@ -42,80 +44,80 @@ public class IssueController {
     public IssueNumberResponseDTO create(@RequestBody IssueRequestDTO issueRequestDTO, @UserId Long userId) {
         logger.debug("이슈 생성");
         logger.debug("issue 요청 확인: {}", issueRequestDTO.toString());
-        return issueService.createIssue(issueRequestDTO, userId);
+        return issueCommandService.createIssue(issueRequestDTO, userId);
     }
 
     @LoginRequired
     @GetMapping("/{issueId}")
     public IssueDetailPageResponseDTO viewDetailPage(@PathVariable Long issueId, @UserId Long userId) {
         logger.debug("이슈 상세 페이지");
-        return issueService.getDetailPage(issueId, userId);
+        return issueQueryService.getDetailPage(issueId, userId);
     }
 
     @PatchMapping("/{issueId}/title")
     public void updateTitle(@PathVariable Long issueId, @RequestBody IssueTitleDTO issueTitleDTO) {
         logger.debug("이슈 제목 수정");
         logger.debug("issue 제목 수정 요청 확인: {}", issueTitleDTO.toString());
-        issueService.updateIssueTitle(issueId, issueTitleDTO);
+        issueCommandService.updateIssueTitle(issueId, issueTitleDTO);
     }
 
     @GetMapping("/{issueId}/assignees")
     public AssigneesResponseDTO viewAssignees(@PathVariable Long issueId) {
         logger.debug("이슈의 담당자 가져오기");
-        return issueService.getAssignees(issueId);
+        return issueQueryService.getAssignees(issueId);
     }
 
     @PatchMapping("/{issueId}/assignees")
     public void updateAssignees(@PathVariable Long issueId, @RequestBody AssigneesToUpdateRequestDTO updateAssigneesRequestDTO) {
         logger.debug("이슈의 담당자 편집");
         logger.debug("이슈의 담당자 편집 요청 확인: {}", updateAssigneesRequestDTO.toString());
-        issueService.updateAssignees(issueId, updateAssigneesRequestDTO);
+        issueCommandService.updateAssignees(issueId, updateAssigneesRequestDTO);
     }
 
     @GetMapping("/{issueId}/labels")
     public LabelsInIssueResponseDTO viewLabels(@PathVariable Long issueId) {
         logger.debug("이슈의 레이블 가져오기");
-        return issueService.getLabels(issueId);
+        return issueQueryService.getLabels(issueId);
     }
 
     @PatchMapping("/{issueId}/labels")
     public void updateLabels(@PathVariable Long issueId, @RequestBody LabelsToUpdateRequestDTO labelsToUpdateRequestDTO) {
         logger.debug("이슈의 레이블 편집");
         logger.debug("이슈의 레이블 편집 요청 확인: {}", labelsToUpdateRequestDTO.toString());
-        issueService.updateLabels(issueId, labelsToUpdateRequestDTO);
+        issueCommandService.updateLabels(issueId, labelsToUpdateRequestDTO);
     }
 
     @GetMapping("/{issueId}/milestones")
     public MilestonesInIssueResponseDTO getMilestones(@PathVariable Long issueId) {
         logger.debug("이슈의 마일스톤 가져오기");
-        return issueService.getMilestones(issueId);
+        return issueQueryService.getMilestones(issueId);
     }
 
     @PatchMapping("/{issueId}/milestones")
     public void updateMilestones(@PathVariable Long issueId, @RequestBody MilestoneToUpdateRequestDTO milestonesToUpdateRequestDTO) {
         logger.debug("이슈의 마일스톤 편집");
         logger.debug("이슈의 마일스톤 편집 요청 확인: {}", milestonesToUpdateRequestDTO.toString());
-        issueService.updateMilestone(issueId, milestonesToUpdateRequestDTO);
+        issueCommandService.updateMilestone(issueId, milestonesToUpdateRequestDTO);
     }
 
     @LoginRequired
     @PostMapping("/{issueId}/comments")
     public CommentDTO createComment(@UserId Long userId, @PathVariable Long issueId, @RequestBody CommentDTO comment) {
         logger.debug("이슈의 코멘트 생성");
-        return issueService.createComment(userId, issueId, comment);
+        return issueCommandService.createComment(userId, issueId, comment);
     }
 
     @LoginRequired
     @PatchMapping("/{issueId}/comments")
     public void updateComment(@UserId Long userId, @PathVariable Long issueId, @RequestBody CommentDTO comment) {
         logger.debug("이슈의 코멘트 편집");
-        issueService.updateComment(userId, issueId, comment);
+        issueCommandService.updateComment(userId, issueId, comment);
     }
 
     @LoginRequired
     @DeleteMapping("/{issueId}/comments/{commentId}")
     public void deleteComment(@UserId Long userId, @PathVariable Long issueId, @PathVariable Long commentId) {
         logger.debug("이슈의 코멘트 삭제");
-        issueService.deleteComment(userId, issueId, commentId);
+        issueCommandService.deleteComment(userId, issueId, commentId);
     }
 }
